Financial_AI_Analyst/requirements.txt
phi
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
streamlit